app/requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
python-jose
passlib[bcrypt]
pandas
openpyxl
python-multipart
authlib
itsdangerous
mcp
requests
twilio
anthropic
